Extract engine components into separate files
