+++ src/engine/physics_system.hpp
@@ -19,6 +19,7 @@
 #include <base/grid.hpp>
 #include <base/spatial_types.hpp>
 #include <data/tile_set.hpp>
+#include <engine/base_components.hpp>
 #include <engine/timing.hpp>
 
 #include <entityx/entityx.h>
@@ -31,8 +32,6 @@ namespace rigel { namespace data { namespace map { class Map; }}}
 
 
 namespace rigel { namespace engine {
-
-using BoundingBox = base::Rect<int>;
 
 
 namespace components {
