+++ src/engine/rendering_system.hpp
@@ -20,9 +20,9 @@
 #include <engine/base_components.hpp>
 #include <engine/map_renderer.hpp>
 #include <engine/timing.hpp>
+#include <engine/visual_components.hpp>
 #include <sdl_utils/texture.hpp>
 
-#include <boost/optional.hpp>
 #include <entityx/entityx.h>
 #include <cstdint>
 #include <utility>
@@ -30,74 +30,6 @@
 
 
 namespace rigel { namespace engine {
-
-namespace components {
-
-struct SpriteFrame {
-  SpriteFrame() = default;
-  SpriteFrame(
-    sdl_utils::NonOwningTexture image,
-    base::Vector drawOffset
-  )
-    : mImage(std::move(image))
-    , mDrawOffset(drawOffset)
-  {
-  }
-
-  sdl_utils::NonOwningTexture mImage;
-  base::Vector mDrawOffset;
-};
-
-
-struct Sprite {
-  std::vector<SpriteFrame> mFrames;
-  int mDrawOrder;
-
-  std::vector<int> mFramesToRender;
-};
-
-
-struct AnimationSequence {
-  AnimationSequence() = default;
-  explicit AnimationSequence(
-    const int delayInTicks,
-    boost::optional<int> endFrame = boost::none
-  )
-    : AnimationSequence(delayInTicks, 0, endFrame)
-  {
-  }
-
-  AnimationSequence(
-    const int delayInTicks,
-    const int startFrame,
-    boost::optional<int> endFrame,
-    const int renderSlot = 0,
-    const bool pingPong = false
-  )
-    : mDelayInTicks(delayInTicks)
-    , mStartFrame(startFrame)
-    , mEndFrame(std::move(endFrame))
-    , mRenderSlot(renderSlot)
-    , mPingPong(pingPong)
-  {
-  }
-
-  int mDelayInTicks = 0;
-  int mStartFrame = 0;
-  boost::optional<int> mEndFrame;
-  int mRenderSlot = 0;
-  bool mPingPong = false;
-
-  TimeStepper mTimeStepper;
-  bool mIsInPingPongLoopBackPhase = false;
-};
-
-
-struct Animated {
-  std::vector<AnimationSequence> mSequences;
-};
-
-}
 
 /** Renders the map and in-game sprites, optionally animating them
  *
